FX-3560 Make modal header and footer sticky
