+++ src/v2/Components/ArtworkFilter/SavedSearch/Components/CreateAlertButton.tsx
@@ -1,9 +1,10 @@
 import React, { useState } from "react"
 import { BellIcon, Button, ButtonProps } from "@artsy/palette"
-import { CreateSavedSearchAlert } from "v2/Components/SavedSearchAlert/CreateSavedSearchAlert"
 import { SavedSearchAttributes } from "../types"
 import { useTracking } from "v2/System"
 import { ActionType, PageOwnerType } from "@artsy/cohesion"
+import { SavedSearchAlertModal } from "v2/Components/SavedSearchAlert/SavedSearchAlertModal"
+import { SavedSearchAlertMutationResult } from "v2/Components/SavedSearchAlert/SavedSearchAlertModel"
 
 interface CreateAlertButtonProps extends ButtonProps {
   savedSearchAttributes: SavedSearchAttributes
@@ -18,7 +19,6 @@ export const CreateAlertButton: React.FC<CreateAlertButtonProps> = ({
 
   const handleOpenForm = () => {
     setVisibleForm(true)
-
     tracking.trackEvent({
       action: ActionType.clickedCreateAlert,
       context_page_owner_type: savedSearchAttributes.type as PageOwnerType,
@@ -26,7 +26,18 @@ export const CreateAlertButton: React.FC<CreateAlertButtonProps> = ({
       context_page_owner_slug: savedSearchAttributes.slug,
     })
   }
-  const handleCloseForm = () => setVisibleForm(false)
+
+  const handleComplete = (result: SavedSearchAlertMutationResult) => {
+    setVisibleForm(false)
+    const trackInfo = {
+      action_type: ActionType.toggledSavedSearch,
+      context_page_owner_type: savedSearchAttributes.type as PageOwnerType,
+      context_page_owner_id: savedSearchAttributes.id,
+      context_page_owner_slug: savedSearchAttributes.slug,
+      saved_search_id: result.id,
+    }
+    tracking.trackEvent(trackInfo)
+  }
 
   return (
     <>
@@ -39,11 +50,12 @@ export const CreateAlertButton: React.FC<CreateAlertButtonProps> = ({
         <BellIcon mr={0.5} color="currentColor" />
         Create an Alert
       </Button>
-      <CreateSavedSearchAlert
-        savedSearchAttributes={savedSearchAttributes}
-        onClose={handleCloseForm}
-        onComplete={handleCloseForm}
+      <SavedSearchAlertModal
         visible={visibleForm}
+        initialValues={{ name: "", email: true, push: false }}
+        savedSearchAttributes={savedSearchAttributes}
+        onClose={() => setVisibleForm(false)}
+        onComplete={handleComplete}
       />
     </>
   )
